randmtzig.c: prevent left-shift of a 1 into sign bit in oct_init_by_entropy().
* randmtzig.c: in expression "word[0] + ... + (word[3]<<24)", cast the latter
  to uint32_t before shifting, preventing (C99) undefined behavior.

+++ liboctave/numeric/randmtzig.c
@@ -268,7 +268,7 @@ oct_init_by_entropy (void)
             unsigned char word[4];
             if (fread (word, 4, 1, urandom) != 1)
               break;
-            entropy[n++] = word[0]+(word[1]<<8)+(word[2]<<16)+(word[3]<<24);
+            entropy[n++] = word[0]+(word[1]<<8)+(word[2]<<16)+((uint32_t)word[3]<<24);
           }
         fclose (urandom);
       }
